Added facebook an gplus oauth. Added templates for not logged users

+++ marketgardenlog/marketgardenlog.py
@@ -7,9 +7,19 @@ from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Garden, Plant, User
 
 import json
+import random
+import string
 
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+from flask import make_response
+import requests
 
 app = Flask(__name__)
+
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
 
 engine = create_engine('sqlite:///marketgardenlog.db')
 Base.metadata.bind = engine
@@ -34,6 +44,8 @@ def gardenPlantsJSON(garden_id):
 @app.route('/gardens/')
 def showGardens():
     gardens = session.query(Garden).order_by(asc(Garden.created))
+    if 'username' not in login_session:
+        return render_template('publicgardens.html', gardens=gardens)
     return render_template('gardens.html', gardens=gardens)
 
 
@@ -41,11 +53,16 @@ def showGardens():
 def showGarden(garden_id):
     garden = session.query(Garden).filter_by(id=garden_id).one()
     plants = session.query(Plant).filter_by(garden_id=garden_id).all()
+    if 'username' not in login_session:
+        return render_template('publicgarden.html',
+                               garden=garden, plants=plants)
     return render_template('garden.html', garden=garden, plants=plants)
 
 
 @app.route('/gardens/<int:garden_id>/edit/', methods=['GET', 'POST'])
 def editGarden(garden_id):
+    if 'username' not in login_session:
+        return redirect('login')
     garden = session.query(Garden).filter_by(id=garden_id).one()
     if request.method == 'POST':
         if request.form['name']:
@@ -62,6 +79,8 @@ def editGarden(garden_id):
 
 @app.route('/gardens/<int:garden_id>/delete/', methods=['GET', 'POST'])
 def deleteGarden(garden_id):
+    if 'username' not in login_session:
+        return redirect('login')
     garden = session.query(Garden).filter_by(id=garden_id).one()
     if request.method == 'POST':
         session.delete(garden)
@@ -74,7 +93,8 @@ def deleteGarden(garden_id):
 
 @app.route('/gardens/new/', methods=['GET', 'POST'])
 def newGarden():
-    login_session['user_id'] = 1
+    if 'username' not in login_session:
+        return redirect('login')
     if request.method == 'POST':
         new_garden = Garden(name=request.form['name'],
                             garden_type=request.form['type'],
@@ -91,7 +111,8 @@ def newGarden():
 
 @app.route('/gardens/<int:garden_id>/newplant/', methods=['GET', 'POST'])
 def newPlant(garden_id):
-    login_session['user_id'] = 1
+    if 'username' not in login_session:
+        return redirect('login')
     if request.method == 'POST':
         new_plant = Plant(name=request.form['name'],
                           plant_type=request.form['type'],
@@ -107,6 +128,8 @@ def newPlant(garden_id):
 @app.route('/gardens/<int:garden_id>/<int:plant_id>/edit/',
            methods=['GET', 'POST'])
 def editPlant(garden_id, plant_id):
+    if 'username' not in login_session:
+        return redirect('login')
     plant = session.query(Plant).filter_by(id=plant_id).one()
     if request.method == 'POST':
         print "Inside POST"
@@ -132,9 +155,34 @@ def editPlant(garden_id, plant_id):
         return render_template('edit_plant.html', plant=plant)
 
 
+# User Helper Functions
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+
 @app.route('/gardens/<int:garden_id>/<int:plant_id>/delete/',
            methods=['GET', 'POST'])
 def deletePlant(garden_id, plant_id):
+    if 'username' not in login_session:
+        return redirect('login')
     plant = session.query(Plant).filter_by(id=plant_id).one()
     if request.method == 'POST':
         session.delete(plant)
@@ -154,9 +202,206 @@ def login():
     return render_template('login.html')
 
 
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_credentials = login_session.get('credentials')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_credentials is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already \
+                                 connected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['credentials'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['provider'] = 'google'
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    # See if a user exists, if it doesn't make a new one
+    user_id = getUserID(data['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: \
+              150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+
+# DISCONNECT - Revoke a current user's token and reset their login_session
+def gdisconnect():
+    # Only disconnect a connected user.
+    credentials = login_session.get('credentials')
+    if credentials is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = credentials
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    return h.request(url, 'GET')[0]
+
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+
+    # Exchange client token for long-lived server-side token
+    app_id = json.loads(open('fb_client_secrets.json',
+                        'r').read())['web']['app_id']
+    app_secret = json.loads(open('fb_client_secrets.json',
+                            'r').read())['web']['app_secret']
+    url = "https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s" % (app_id, app_secret, access_token)  # NOQA
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+    # Strip expire tag from access token
+    token = result.split("&")[0]
+
+    # Get user info
+    url = 'https://graph.facebook.com/v2.4/me?%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+    login_session['credentials'] = access_token
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.2/me/picture?%s&redirect=0&height=200&width=200' % token  # NOQA
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # See if a user exists, if it doesn't make a new one
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: \
+              150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    access_token = login_session.get('credentials')
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id, access_token)  # NOQA
+    # url = 'https://graph.facebook.com/%s/permissions' % facebook_id
+    h = httplib2.Http()
+    return h.request(url, 'DELETE')[1]
+
+
 @app.route('/logout')
 def logout():
-    return redirect('/')
+    print login_session
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            result = gdisconnect()
+            del login_session['gplus_id']
+        if login_session['provider'] == 'facebook':
+            result = fbdisconnect()
+            del login_session['facebook_id']
+        print result
+        del login_session['credentials']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        del login_session['state']
+        flash("You have successfully been logged out.")
+        print login_session
+        return redirect(url_for('showGardens'))
+
+    else:
+        flash("You were not logged in")
+        return redirect(url_for('showGardens'))
 
 
 if __name__ == '__main__':
